Address all PR reviewer feedback
- Updated documentation link to AWS SDK v2
- Changed context.TODO() to context.Background() for initialization
- Made StorageCapacity int32 throughout codebase
- Made newSizeGiB parameter int32 for resize operations
- Used ErrorMessage() method instead of direct field access
- Updated AWS SDK packages to latest versions (FSx v1.55.0)
- Updated Go version to 1.24.0
- Removed unnecessary log statement for consistency
- Changed getter functions to return by value instead of pointer
- Refactored getTagsForResource to accept context parameter
- Added client-side validation for delete parameters
- Replaced fsxtypes alias with types for consistency
- Restored EC2 metadata availability check using GetInstanceIdentityDocument
- Added back missing test functions updated for AWS SDK v2
- Fixed all compilation errors and test failures

+++ pkg/driver/controller.go
@@ -218,7 +218,7 @@ func (d *controllerService) CreateVolume(ctx context.Context, req *csi.CreateVol
 
 		return &csi.CreateVolumeResponse{
 			Volume: &csi.Volume{
-				CapacityBytes: util.GiBToBytes(fs.StorageCapacity),
+				CapacityBytes: util.GiBToBytes(int64(fs.StorageCapacity)),
 				VolumeId:      fs.FileSystemId,
 				VolumeContext: map[string]string{
 					volumeContextDnsName:      fs.DnsName,
@@ -595,16 +595,16 @@ func (d *controllerService) ControllerExpandVolume(ctx context.Context, req *csi
 			return nil, status.Errorf(codes.Internal, "Could not get filesystem with ID %q: %v", volumeID, err)
 		}
 
-		if newCapacity <= fs.StorageCapacity {
+		if newCapacity <= int64(fs.StorageCapacity) {
 			// Current capacity is sufficient to satisfy the request
 			klog.V(4).InfoS("ControllerExpandVolume: current filesystem capacity matches or exceeds requested storage capacity, returning with success", "currentStorageCapacityGiB", fs.StorageCapacity, "requestedStorageCapacityGiB", newCapacity)
 			return &csi.ControllerExpandVolumeResponse{
-				CapacityBytes:         util.GiBToBytes(fs.StorageCapacity),
+				CapacityBytes:         util.GiBToBytes(int64(fs.StorageCapacity)),
 				NodeExpansionRequired: false,
 			}, nil
 		}
 
-		finalCapacity, err := d.cloud.ResizeFileSystem(ctx, volumeID, newCapacity)
+		finalCapacity, err := d.cloud.ResizeFileSystem(ctx, volumeID, int32(newCapacity))
 		if err != nil {
 			return nil, status.Errorf(codes.Internal, "resize failed: %v", err)
 		}
@@ -615,7 +615,7 @@ func (d *controllerService) ControllerExpandVolume(ctx context.Context, req *csi
 		}
 
 		return &csi.ControllerExpandVolumeResponse{
-			CapacityBytes:         util.GiBToBytes(*finalCapacity),
+			CapacityBytes:         util.GiBToBytes(int64(*finalCapacity)),
 			NodeExpansionRequired: false,
 		}, nil
 	}
